try to set PYTHONPATH from running environment
This doesn't work as well as I'd like because I start emacsd with
systemd's user sessions, and that doesn't read the environment variables
from zsh when I start it - it only detects what's in the unit file.

So there are possible workarounds, but none of them are really this
one (unless I happen to be running on a system that doesn't use systemd
user sessions for initialization).