move ,B binding to my-leader-keys.el
I used to put leader key bindings in my-buffers.el. But I don't like
doing that.

Reasoning: I have a lot of leader key bindings. If they are in different
files, there's a good shot that a future one will replace an earlier
one, and my Emacs config won't make any sense (and it will be hard to edit).