core: try to make fonts work
Since I'm using Emacs in daemon mode, it's a big wonky trying to get
fonts to load the first time over. It needs to detect whether we're
running in a graphical frame, and then change the font.

Since when Emacs starts, it's not in a graphical frame, it doesn't
change the font.

This change doesn't work, but it *does* make things slightly easier on me.