place .cache in the repository
Having this present in the repo fixes some new-clone problems
(related: #1 on GitHub). Namely, if you do not have a .cache
folder when you clone the repo, it isn't created automatically,
and smex tries to look for a history/item cache in that folder.

This means M-x fails on new clones, which is pretty horrible.