redefine py-help-at-point to do nothing
py-help-at-point automaically jumps to the definition of whatever it is
called on. Since eldoc attempts to call this function to get a help
message for the modeline, it jumps your cursor around arbitrarily in
Python files.

Which sucks. Should probably file a bug upstream, but I'm not sure whose
fault it is yet ...